replace: bump F-Stack embedded DPDK from 23.11.5 to 24.11.6 LTS (tree replace)
Pristine integration of upstream DPDK 24.11.6 LTS into f-stack/dpdk/. F-Stack
local patches (29c7d5835 + 5f3768c63 + 62f1c34df + 92718178b) are NOT in
this commit; they will be re-applied in the immediately-following 'port:'
commit per plan.md §4.4 (single merged 'port:' commit per DP-A8). Patch list
updated 3 -> 4 after user 2026-06-09 14:50 KNI feedback uncovered patch-scout's
missed identification of 29c7d5835.

Versions:
  upstream: VERSION=24.11.6, ABI_VERSION=25.0
  previous: VERSION=23.11.5, ABI_VERSION=24.0

lib/ count: 57 -> 59 (+argparse +ptr_compress, both opt-in, F-Stack 0 ref).

Backed up old tree to f-stack/dpdk.bak-23.11.5/ (kept until M5 PASS).
F-Stack patches saved as git format-patch under
/data/workspace/dpdk-stable-24.11.6/f-stack-lib/patches/.

Build verification (M2 acceptance):
  meson setup build:  exit=0 / 749 build targets / 0 errors / 2 warnings
                      (only missing optional PMD deps: mvep, af_xdp, etc.)
  ninja -C build:     exit=0 / 0 errors / 0 warnings / 2m14s build
                      (vs 23.11.5 baseline 4 warnings; cleaner)
  build size:         653 MB (vs 23.11.5 623 MB, +4.8% within NFR-D-3 ±30%)
  librte_eal.a:       875438 bytes (vs 835328, +4.8%)
  librte_ethdev.a:    874896 bytes (vs 825506, +6.0%)
  librte_argparse.a:  16624 bytes (NEW, F-Stack 0 ref - confirmed lean cost)

KNI / igb_uio status verified absent in 24.11.6 upstream:
  ls dpdk/lib/kni                 -> No such file or directory
  ls dpdk/kernel/linux/kni        -> No such file or directory
  ls dpdk/kernel/linux/igb_uio    -> No such file or directory
(Will be re-introduced (igb_uio) and kept absent (kni) by next 'port:' commit
 per spec 02 §3.6 + 04 §4.2.0/4.2.1)

Local commit only; no push.

+++ dpdk/drivers/net/i40e/base/i40e_dcb.c
@@ -906,7 +906,7 @@ enum i40e_status_code i40e_init_dcb(struct i40e_hw *hw, bool enable_mib_change)
 		return I40E_ERR_NOT_READY;
 
 	/* Get the LLDP AdminStatus for the current port */
-	adminstatus = lldp_cfg.adminstatus >> (hw->port * 4);
+	adminstatus = (u8)(lldp_cfg.adminstatus >> (hw->port * 4));
 	adminstatus &= 0xF;
 
 	/* LLDP agent disabled */
@@ -1307,7 +1307,6 @@ enum i40e_status_code i40e_dcb_config_to_lldp(u8 *lldpmib, u16 *miblen,
 	*miblen = offset;
 	return ret;
 }
-
 
 /**
  * _i40e_read_lldp_cfg - generic read of LLDP Configuration data from NVM
